Moved the controller and DAO beans from the applicationContext to the springServlet file so that the servlet mappings to the controller could be fulfilled.
Setup servlet mapping for the spring DispatcherSerlvet and springContext
in the web.xml.

Added a simple get method to the controller class.
Added an addPerson and getPerson methods to the controller class to add
a person to our dao and retrieve a person.

Added unit tests for the controller behaviour, as well as debug print
lines and a new toString for the Person domain object.